requirements.txt
pillow
flask
pyvips
einops
torch
transformers
accelerate
